명산정보페이지 수정완료

+++ routes/mypage/mypage.js
@@ -20,18 +20,7 @@ router.use((req, res, next) => {
 //프로필 사진 읽기
 router.get("/", async (req, res, next) => {
   try {
-    // const getLike = await Club.findAll({
-    //   include: {
-    //     model: User,
-    //     attributes: ["id"],
-    //   },
-    // });
-
-    console.log("////////////////////////////////////////////////////");
-    // console.log(getLike);
-    console.log("////////////////////////////////////////////////////");
     const getImage = await User.findOne({
-      // where: { id: req.user.id }, // 여기서 id 에러창뜨고 로그인됨
       where: { id: `${req.user.id}` }, // 문자로 바꿈. 아마도. 암튼 개선함
     });
     res.render("mypage/mypage", {
@@ -46,39 +35,16 @@ router.get("/", async (req, res, next) => {
 
 //회원계정 삭제 //가입되지않은 회원이라뜨는데db에는 남아있음..
 router.get("/:id/delete", async (req, res, next) => {
-  try {
-    const getID = await User.findOne({ where: { id: `${req.user.id}` } });
-    await getID.destroy({ where: { id: `${req.user.id}` } });
+  if (confirm("계정을 삭제하시겠습니까?")) {
+    try {
+      const getID = await User.findOne({ where: { id: `${req.user.id}` } });
+      await getID.destroy({ where: { id: `${req.user.id}` } });
 
-    res.redirect("/");
-  } catch (err) {
-    console.error(err);
-    next(err);
+      res.redirect("/");
+    } catch (err) {
+      console.error(err);
+      next(err);
+    }
   }
 });
-
-// //회원계정 삭제 //가입되지않은 회원이라뜨는데db에는 남아있음..
-// router.get('/:id/delete',async (req,res,next)=>{
-
-//    try{
-//      const getID = await User.findOne({ where:{id: `${req.user.id}`}});
-//      await getID.destroy({ id: req.user.id});
-
-//      res.redirect('/');
-//    }.then(()=>{
-//      req.session.destroy((err)=>{
-//        if(err){
-//          res.status(400).send('you are currently not logined');
-//        }else {
-//          res.status(200).send('안전하게 퇼퇴');
-//        }
-//      });
-//    }).catch((err)=>{
-//      res.status(500).send('err');
-//    })
-// });
-
-// router.route('/:id')
-// .delete(User);
-
 module.exports = router;
